Decompose MongoDriverSpecialTest into themed classes.
Split the 1657-line catch-all into five purpose-named classes: ReconnectTest
(outage and reconnect behavior), RefurbishTest, SchemaCompatibilityTest (tolerance
of state written by a different type version), MetadataTest (manifest and document
metadata, folding in the ManifestTest), and SpuriousActionTest (the driver takes no
spurious actions). Hoist rootDocumentsFilter into AbstractMongoDriverTest, factor
TestParameters.standardDriverSettings, and delete the obsolete warmStart_stateMatches
and listing_stateMatches tests. No test scenarios are lost.
